Move code used to render text nodes from `Node` into a separate React component.
